requirements.txt
playwright==1.49.1
pytest==8.3.4
pytest-playwright==0.6.2
allure-pytest==2.13.5
python-dotenv==1.0.1
faker==33.0.0
pytest-xdist==3.5.0          # параллельный запуск
pytest-rerunfailures==14.0   # перезапуск упавших
requests==2.32.3             # используется в utils/api_client.py (cleanup / API-хелперы)